Move types `Sort` and `SearchParams` to more general `PaginationTypes` file.
